@@ -1,6 +1,6 @@
 module github.com/free5gc/udr
 
-go 1.17
+go 1.21
 
 require (
 	github.com/asaskevich/govalidator v0.0.0-20210307081110-f21760c49a8d
